requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
language-tool-python==2.8
MarkupSafe==2.1.5
nltk==3.8.1
pyenchant==3.2.2
regex==2024.4.16
requests==2.31.0
tqdm==4.66.2
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
